+++ app/auth/signup/SignUp.tsx
@@ -8,9 +8,9 @@ import Link from "next/link";
 import { Eye, EyeOff } from "lucide-react";
 import Image from "next/image";
 import Spinner from "@/app/components/Spinner";
-import Toast from "@/app/components/Toast";
-import { useToast } from "@/hooks/use-toast";
+// import Toast from "@/app/components/Toast";
 import withAuthRedirect from "@/app/hoc/withAuthRedirect";
+import { useToast } from "@/app/hooks/useToast";
 
 
 // Define TypeScript interface for form values
