+++ app/auth/signup/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import Signup from './SignUp'
 import Navbar from '@/app/components/Navbar'
-import Toast from '@/app/components/Toast'
+import Toast from '@/app/components/Toaster'
 
 
 const Authendication = () => {
