+++ app/components/Toaster.tsx
@@ -1,18 +1,10 @@
 "use client"
-
-import { useToast } from "@/hooks/use-toast"
-import {
-  Toast,
-  ToastClose,
-  ToastDescription,
-  ToastProvider,
-  ToastTitle,
-  ToastViewport,
-} from "@/components/ui/toast"
+import { useToast } from "../hooks/useToast";
+import { Toast, ToastClose, ToastDescription, ToastProvider, ToastTitle, ToastViewport } from "./toast"
+import React from "react";
 
 export function Toaster() {
   const { toasts } = useToast()
-
   return (
     <ToastProvider>
       {toasts.map(function ({ id, title, description, action, ...props }) {
@@ -29,7 +21,7 @@ export function Toaster() {
           </Toast>
         )
       })}
-      <ToastViewport/>
+      <ToastViewport />
     </ToastProvider>
   )
 }
