+++ app/hooks/useToast.ts
@@ -1,75 +1,191 @@
-import { useState, useCallback } from "react";
+"use client"
 
-type ToastType = "success" | "error" | "info" | "warning";
+// Inspired by react-hot-toast library
+import * as React from "react"
+import { ToastActionElement, ToastProps } from "../components/toast"
 
-interface ToastOptions {
-  type: ToastType;
-  description: string;
-  duration?: number; // Optional duration for the toast
-  position?: "top-right" | "top-left" | "bottom-right" | "bottom-left"; // Optional position, defaults to top-right
+
+const TOAST_LIMIT = 1
+const TOAST_REMOVE_DELAY = 1000000
+
+type ToasterToast = ToastProps & {
+  id: string
+  title?: React.ReactNode
+  description?: React.ReactNode
+  action?: ToastActionElement
 }
 
-export const useToast = () => {
-  const [toast, setToast] = useState<ToastOptions | null>(null);
+const actionTypes = {
+  ADD_TOAST: "ADD_TOAST",
+  UPDATE_TOAST: "UPDATE_TOAST",
+  DISMISS_TOAST: "DISMISS_TOAST",
+  REMOVE_TOAST: "REMOVE_TOAST",
+} as const
 
-  const showToast = useCallback((options: ToastOptions) => {
-    setToast(options);
-    if (options.duration) {
-      setTimeout(() => {
-        setToast(null); // Hide toast after the duration
-      }, options.duration);
+let count = 0
+
+function genId() {
+  count = (count + 1) % Number.MAX_SAFE_INTEGER
+  return count.toString()
+}
+
+type ActionType = typeof actionTypes
+
+type Action =
+  | {
+      type: ActionType["ADD_TOAST"]
+      toast: ToasterToast
     }
-  }, []);
+  | {
+      type: ActionType["UPDATE_TOAST"]
+      toast: Partial<ToasterToast>
+    }
+  | {
+      type: ActionType["DISMISS_TOAST"]
+      toastId?: ToasterToast["id"]
+    }
+  | {
+      type: ActionType["REMOVE_TOAST"]
+      toastId?: ToasterToast["id"]
+    }
 
-  const hideToast = useCallback(() => {
-    setToast(null); // Hide toast
-  }, []);
+interface State {
+  toasts: ToasterToast[]
+}
 
-  // Toast methods for different types
-  const toastSuccess = (message: string, duration?: number, position: "top-right" | "top-left" | "bottom-right" | "bottom-left" = "top-right") => {
-    showToast({
-      type: "success",
-      description: message,
-      duration,
-      position,
-    });
-  };
+const toastTimeouts = new Map<string, ReturnType<typeof setTimeout>>()
 
-  const toastError = (message: string, duration?: number, position: "top-right" | "top-left" | "bottom-right" | "bottom-left" = "top-right") => {
-    showToast({
-      type: "error",
-      description: message,
-      duration,
-      position,
-    });
-  };
+const addToRemoveQueue = (toastId: string) => {
+  if (toastTimeouts.has(toastId)) {
+    return
+  }
 
-  const toastInfo = (message: string, duration?: number, position: "top-right" | "top-left" | "bottom-right" | "bottom-left" = "top-right") => {
-    showToast({
-      type: "info",
-      description: message,
-      duration,
-      position,
-    });
-  };
+  const timeout = setTimeout(() => {
+    toastTimeouts.delete(toastId)
+    dispatch({
+      type: "REMOVE_TOAST",
+      toastId: toastId,
+    })
+  }, TOAST_REMOVE_DELAY)
 
-  const toastWarning = (message: string, duration?: number, position: "top-right" | "top-left" | "bottom-right" | "bottom-left" = "top-right") => {
-    showToast({
-      type: "warning",
-      description: message,
-      duration,
-      position,
-    });
-  };
+  toastTimeouts.set(toastId, timeout)
+}
+
+export const reducer = (state: State, action: Action): State => {
+  switch (action.type) {
+    case "ADD_TOAST":
+      return {
+        ...state,
+        toasts: [action.toast, ...state.toasts].slice(0, TOAST_LIMIT),
+      }
+
+    case "UPDATE_TOAST":
+      return {
+        ...state,
+        toasts: state.toasts.map((t) =>
+          t.id === action.toast.id ? { ...t, ...action.toast } : t
+        ),
+      }
+
+    case "DISMISS_TOAST": {
+      const { toastId } = action
+
+      // ! Side effects ! - This could be extracted into a dismissToast() action,
+      // but I'll keep it here for simplicity
+      if (toastId) {
+        addToRemoveQueue(toastId)
+      } else {
+        state.toasts.forEach((toast) => {
+          addToRemoveQueue(toast.id)
+        })
+      }
+
+      return {
+        ...state,
+        toasts: state.toasts.map((t) =>
+          t.id === toastId || toastId === undefined
+            ? {
+                ...t,
+                open: false,
+              }
+            : t
+        ),
+      }
+    }
+    case "REMOVE_TOAST":
+      if (action.toastId === undefined) {
+        return {
+          ...state,
+          toasts: [],
+        }
+      }
+      return {
+        ...state,
+        toasts: state.toasts.filter((t) => t.id !== action.toastId),
+      }
+  }
+}
+
+const listeners: Array<(state: State) => void> = []
+
+let memoryState: State = { toasts: [] }
+
+function dispatch(action: Action) {
+  memoryState = reducer(memoryState, action)
+  listeners.forEach((listener) => {
+    listener(memoryState)
+  })
+}
+
+type Toast = Omit<ToasterToast, "id">
+
+function toast({ ...props }: Toast) {
+  const id = genId()
+
+  const update = (props: ToasterToast) =>
+    dispatch({
+      type: "UPDATE_TOAST",
+      toast: { ...props, id },
+    })
+  const dismiss = () => dispatch({ type: "DISMISS_TOAST", toastId: id })
+
+  dispatch({
+    type: "ADD_TOAST",
+    toast: {
+      ...props,
+      id,
+      open: true,
+      onOpenChange: (open) => {
+        if (!open) dismiss()
+      },
+    },
+  })
 
   return {
-    toast: {
-      success: toastSuccess,
-      error: toastError,
-      info: toastInfo,
-      warning: toastWarning,
-    },
-    hideToast,
-    toastContent: toast, // The current toast message and options
-  };
-};
+    id: id,
+    dismiss,
+    update,
+  }
+}
+
+function useToast() {
+  const [state, setState] = React.useState<State>(memoryState)
+
+  React.useEffect(() => {
+    listeners.push(setState)
+    return () => {
+      const index = listeners.indexOf(setState)
+      if (index > -1) {
+        listeners.splice(index, 1)
+      }
+    }
+  }, [state])
+
+  return {
+    ...state,
+    toast,
+    dismiss: (toastId?: string) => dispatch({ type: "DISMISS_TOAST", toastId }),
+  }
+}
+
+export { useToast, toast }
