+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@radix-ui/react-dropdown-menu": "^2.1.6",
     "@radix-ui/react-toast": "^1.2.6",
     "@supabase/supabase-js": "^2.43.4",
     "@vercel/analytics": "^1.3.1",
